ldsk12_beta v2.1.1
- CLA view display
- Lesson Task Relation API Change
- bugs fix

+++ database/migrations/2020_12_22_154519_add_column_lesson_task_relation_starttime.php
@@ -16,7 +16,7 @@ class AddColumnLessonTaskRelationStarttime extends Migration
         //
         Schema::table('lesson_task_relation', function($table)
         {
-            $table->smallInteger('starttime')->nullable();
+            $table->bigInteger('starttime')->nullable();
         });
     }
